Fix flaky test (#52088)
